static.css the style="" attributes and fix download.html in large fonts.
I've spent a lot of time debugging the style="" attributes in the HTML
which over-rided all my changes to the stylesheet. I got rid of them
and replaced them with saner ones in the static.css.

I also fixed the fact that when resizing fonts the right column sometimes
overflowed to be placed below the left one.


git-svn-id: http://svn.perlide.org/padre/trunk/template.padre.perlide.org@7595 bc93166f-e056-4eb9-8c59-0a0840647e1b